implement Exception Handler

+++ src/main/java/com/livares/intern/Service/implementation/CategoryServiceImpl.java
@@ -1,4 +1,4 @@
-package com.livares.intern.Service;
+package com.livares.intern.service.implementation;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,8 +6,11 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.livares.intern.Model.Category;
-import com.livares.intern.Repository.CategoryRepository;
+import com.livares.intern.exception.CustomException;
+import com.livares.intern.exception.ErrorCode;
+import com.livares.intern.model.Category;
+import com.livares.intern.repository.CategoryRepository;
+import com.livares.intern.service.CategoryService;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
@@ -26,9 +29,14 @@ public class CategoryServiceImpl implements CategoryService {
     public Category saveCategory(Category category) {
         return categoryRepository.save(category);
     }
-
+    
+    @Override
     public void deleteCategory(Long id) {
+        Optional<Category> categoryid = categoryRepository.findById(id);
+       if(categoryid.isPresent()){
         categoryRepository.deleteById(id);
+       }else {
+           throw new CustomException(ErrorCode.NOT_FOUND,"This "+id+" not found");
+       }
     }
-	
 }
